+++ packages/app/src/runner/injectBundle.ts
@@ -1,3 +1,4 @@
+import { initializeEventManager } from '.'
 import { initializeMobxStore, useAutStore } from '../store'
 
 export async function injectBundle () {
@@ -37,6 +38,10 @@ export async function injectBundle () {
       autStore.updateDimensions(config.viewportWidth, config.viewportHeight)
 
       window.UnifiedRunner.config = config
+
+      // window.UnifiedRunner exists now, since the Webpack bundle with
+      // the UnifiedRunner namespace was injected.
+      initializeEventManager(window.UnifiedRunner)
 
       window.UnifiedRunner.MobX.runInAction(() => {
         const store = initializeMobxStore(window.UnifiedRunner.config.testingType)