+++ packages/server/lib/util/specs.ts
@@ -1,5 +1,4 @@
 import _ from 'lodash'
-import R from 'ramda'
 import la from 'lazy-ass'
 import path from 'path'
 import check from 'check-more-types'
@@ -184,15 +183,13 @@ function findSpecsOfType (searchFolder: string, commonSearchOptions: CommonSearc
   return Bluebird.mapSeries(testFilesPatterns, findOnePattern).then(_.flatten)
 }
 
-const setTestType = (testType: Cypress.CypressSpecType) => R.map(R.set(R.lensProp('specType'), testType))
-
 const findIntegrationSpecs = (searchFolder: string | undefined, commonSearchOptions: CommonSearchOptions, specPattern: string | undefined) => {
   if (!searchFolder) {
     return []
   }
 
   return findSpecsOfType(searchFolder, commonSearchOptions, specPattern)
-  .then(setTestType(SPEC_TYPES.INTEGRATION))
+  .then((val) => val.map((s) => ({ ...s, specType: SPEC_TYPES.INTEGRATION })))
 }
 
 const findComponentSpecs = (searchFolder: string | undefined | false, integrationFolder: string | undefined | false, commonSearchOptions: CommonSearchOptions, specPattern: string | undefined) => {
@@ -208,7 +205,7 @@ const findComponentSpecs = (searchFolder: string | undefined | false, integratio
   ]
 
   return findSpecsOfType(searchFolder, { ...commonSearchOptions, ignoreTestFiles }, specPattern)
-  .then(setTestType(SPEC_TYPES.COMPONENT))
+  .then((val) => val.map((s) => ({ ...s, specType: SPEC_TYPES.COMPONENT })))
 }
 
 const printFoundSpecs = (foundSpecs: Cypress.Spec[]) => {